Add Control.Effect.Resource, which provides a pseudo-effect
Providing Resource as a bona fide effect is dangerous, as it is unsound
to lift Resource through arbitrary handlers: handlers that run the same
computation several times could do unexpected things, and handlers that
introduce new failure modes (such as Error) could abort the computation
without the cleanup action being run!

To avoid this, Resource is implemented as a combination of a Mask effect
and a MonadUnwind mtl-style typeclass. This requires a MonadUnwind
instance for every handler, but this isn’t as bad as it sounds, as most
custom handlers can be implemented with HandlerT, and the MonadUnwind
instance for HandlerT only needs to be defined exactly once, in this
library. Hopefully, the tradeoff pans out.

Also, as a part of these changes, eff is now compiled with
-flate-specialise, which turns out to be important for performance, and
a few bits of internal machinery are moved around to cooperate with the
specializer a little bit better (though after further investigation,
some of them seem probably not strictly necessary). Unfortunately, the
specializer is not able to fully specialize programs to the degree
desired on GHC 8.8 and below, but testing on GHC 8.10 confirms that
specialization is working as intended. With -flate-specialise,
specialization on GHC 8.6/8.8 seems okay, but some things are not
specialized that should be. More investigation is necessary to determine
whether or not performance is acceptable on GHC versions prior to 8.10.