Add an additional Writer law
This law relates all three methods together to help pin down the meaning
of `listen (e1 <|> e2)`, where (<|>) is handled by NonDetT, more
definitively.

This developed from a conversation with Rob Rix, who suggested that
`listen (tell a <|> tell b)` ought to be equivalent to
`((a <> b),) <$> (tell a <|> tell b)`, not `listen (tell a) <|>
listen (tell b)` when `Writer` has global state semantics with respect
to `Alternative`. I disagree, and this law captures why the former
behavior is not allowed.