Add `control0` and generalize `handle` to accept an exit handler
This allows Coroutine to be given a neater implementation, where Status
allows installing an entirely different handler on each iteration if
that is desired.

The main downside to this change is that it makes `handle` and `Handle`
more complicated, since it needs a little extra bookkeeping to track the
types for the exit handler. Still, this is the advanced API anyway, so
maybe it’s okay. If this turns out to be too confusing, it can
theoretically be relegated to an “extra advanced” API.